Discriminator Threshold
Simple discriminator threshold added to the DAQ that checks on a
photon-by-photon basis to see if the deposited charge is high enough to
allow the PMT to contribute to the global trigger. This does not look at
the total waveform, and thus does not correctly work with photon pileup
(they are treated as statistical instead of correlated).